[refactor] slightly better components organization
