+++ src/features/timetable/Timetable.tsx
@@ -1,10 +1,9 @@
 import { type FC, useCallback, useEffect, useMemo, useState } from 'react';
-import TimetableCell from './components/TimetableCell';
-import TimetableListItem from './components/TimetableListItem';
+import TimetableLesson from './components/TimetableLesson';
 import useWindowDimensions from '@/hooks/useWindowDimensions';
 import { lessonsTimes, unique } from '@/utils/timetable';
 import { getCurrentUADate, stringToDate } from '@/utils/date';
-import { TABLET_SCREEN_BREAKPOINT } from '@/utils/constants';
+import { DEVELOP, TABLET_SCREEN_BREAKPOINT } from '@/utils/constants';
 import type { TimetableItem } from '@/types/timetable';
 import { classes } from '@/styles/utils';
 import styles from './Timetable.module.scss';
@@ -13,12 +12,12 @@ type OwnProps = {
   timetable: TimetableItem[]
   isSecondSubgroup: boolean
   isSecondWeek: boolean
-  cellSubgroup?: boolean
+  hasCellSubgroups?: boolean
 };
 
 const MINUTE = 60 * 1000;
 
-const Timetable: FC<OwnProps> = ({ timetable, isSecondSubgroup, isSecondWeek, cellSubgroup }) => {
+const Timetable: FC<OwnProps> = ({ timetable, isSecondSubgroup, isSecondWeek, hasCellSubgroups }) => {
   const { width } = useWindowDimensions();
   const isMobile = width < TABLET_SCREEN_BREAKPOINT;
   const maxLessonNumber = useMemo(() =>
@@ -59,15 +58,15 @@ const Timetable: FC<OwnProps> = ({ timetable, isSecondSubgroup, isSecondWeek, ce
     const result = timetable.filter(item =>
       item.day === day &&
       item.number === number &&
-      (cellSubgroup || item.isSecondSubgroup === isSecondSubgroup || forBothSubgroups(item)) &&
+      (hasCellSubgroups || item.isSecondSubgroup === isSecondSubgroup || forBothSubgroups(item)) &&
       (item.isSecondWeek === isSecondWeek || forAllWeeks(item))
     );
     return result.length === 0 ? null : unique(result);
-  }, [timetable, isSecondSubgroup, isSecondWeek, cellSubgroup]);
+  }, [timetable, isSecondSubgroup, isSecondWeek, hasCellSubgroups]);
 
   const tableContent = useMemo(() => {
     if (isMobile) return null;
-    console.log('Running scary useMemo');
+    if (DEVELOP) console.log('Running scary useMemo');
     const table = timetableLessonsTimes.map((time, i) =>
       <tr key={time.start}>{
         days.map((day, j) =>
@@ -77,39 +76,47 @@ const Timetable: FC<OwnProps> = ({ timetable, isSecondSubgroup, isSecondWeek, ce
                 <span className={classes(styles.metadata, styles.number)}>{i + 1}</span>
                 <span className={classes(styles.metadata, styles.end)}>{time.end}</span>
               </th>
-            : <TimetableCell
+            : <TimetableLesson
+                isListView={false}
                 isAfterEmpty={i !== 0 && getLessonsByDayAndTime(i, j) === null}
                 lessons={getLessonsByDayAndTime(i + 1, j)}
                 active={currentLessonNumber === i && currentDay === j}
                 key={time.start + day}
-                cellSubgroup={cellSubgroup}
+                cellSubgroup={hasCellSubgroups}
               />
         )
       }</tr>
     );
     return table;
-  }, [timetableLessonsTimes, timetable, days, getLessonsByDayAndTime, currentLessonNumber, currentDay, cellSubgroup]);
+  }, [
+    isMobile, timetableLessonsTimes, timetable, days,
+    getLessonsByDayAndTime, currentLessonNumber, currentDay, hasCellSubgroups
+  ]);
 
   const listsContent = useMemo(() => {
     if (!isMobile) return null;
-    console.log('Running scary useMemo');
+    if (DEVELOP) console.log('Running scary useMemo');
     const lists = days.filter(Boolean).map((day, i) =>
       <div key={i} className={styles.list}>
         <h3>{day}</h3>
         <ol className={styles.list}>{
           timetableLessonsTimes.map((time, j) =>
-            <TimetableListItem
+            <TimetableLesson
+              isListView={true}
               lessons={getLessonsByDayAndTime(j + 1, i + 1)}
               active={currentLessonNumber === j && currentDay === i + 1}
               key={time.start + day}
-              cellSubgroup={cellSubgroup}
+              cellSubgroup={hasCellSubgroups}
             />
           )
         }</ol>
       </div>
     );
     return lists;
-  }, [timetableLessonsTimes, timetable, days, getLessonsByDayAndTime, currentLessonNumber, currentDay, cellSubgroup]);
+  }, [
+    isMobile, timetableLessonsTimes, timetable, days,
+    getLessonsByDayAndTime, currentLessonNumber, currentDay, hasCellSubgroups
+  ]);
 
   return (
     isMobile
