+++ src/features/timetable/components/TimetableLesson.tsx
@@ -1,40 +1,88 @@
-import { type FC, memo } from 'react';
-import TimetableLink from '../ui/TimetableLink';
-import { classes } from '@/styles/utils';
+import { type FC, memo, type ReactNode } from 'react';
+import TimetableLessonInfo from './TimetableLessonInfo';
+import { useDelayedProp } from '@/hooks/useDelayedProp';
+import { generateId, lessonsTimes } from '@/utils/timetable';
 import styles from './TimetableCell.module.scss';
+import { classes } from '@/styles/utils';
 import type { TimetableItem } from '@/types/timetable';
 
 type OwnProps = {
-  lesson: TimetableItem
-  isMerged?: boolean
+  lessons: TimetableItem[] | null
+  active: boolean
+  isListView: boolean
   cellSubgroup?: boolean
+  isAfterEmpty?: boolean
 };
 
-const TimetableLesson: FC<OwnProps> = ({ lesson, cellSubgroup, isMerged }) => {
-  const isForBothSubgroups = lesson.isFirstSubgroup && lesson.isSecondSubgroup;
+const ANIMATION_DURATION = 300;
+
+const colors: Record<TimetableItem['type'], string> = {
+  lection: 'var(--lect-clr)',
+  practical: 'var(--pract-clr)',
+  lab: 'var(--lab-clr)',
+  consultation: 'var(--consult-clr)'
+};
+
+function createStyleAttrFromLessons (lessons: TimetableItem[]) {
+  const propertyObject = lessons.reduce<Record<string, string>>((acc, lesson, index) => {
+    acc[`--color-${index + 1}`] = colors[lesson.type];
+    return acc;
+  }, {});
+  return propertyObject;
+}
+
+const TimetableLesson: FC<OwnProps> = ({ isListView, lessons, active, cellSubgroup, isAfterEmpty }) => {
+  const [innerLessons, shouldAppear] = useDelayedProp(lessons, ANIMATION_DURATION);
+  const isMerged = Boolean(innerLessons && innerLessons.length > 1);
 
   return (
     <>
-      <div className={
-          classes(styles.cell, isMerged && styles.merged, !lesson ? styles.hide : styles.show)
-        }>
-        <div className={styles.info}>
-          {cellSubgroup && !isForBothSubgroups &&
-            <span className={styles.subgroup}>
-              {lesson.isSecondSubgroup ? 'II' : 'I'} підгрупа
-            </span>}
-          <span className={styles.title}>
-            <h2 className={styles.name}>{lesson.subject.replace('`', '’')}</h2>
-            {lesson.lecturer.trim().replace(/,$/, '')}
-          </span>
-          <span className={styles.extra}>
-            {lesson.location.replaceAll(/,./g, '').trim()}
-            <TimetableLink urls={lesson.urls} type={lesson.type} />
-          </span>
-        </div>
-      </div>
+      {innerLessons !== null
+        ? (
+           <LessonContainer
+            isList={isListView}
+            className={classes(
+              styles['cell-container'],
+              active && styles.active,
+              styles[`color-${innerLessons?.length}`],
+              !shouldAppear && !lessons ? styles.hide : styles.show
+            )}
+            style={createStyleAttrFromLessons(innerLessons ?? [])}
+           >
+              <div
+                className={classes(styles.spacer, innerLessons?.[0]?.type)}
+                data-time={
+                  `${lessonsTimes[innerLessons?.[0]?.number ?? 0 - 1]?.start}
+                  -${lessonsTimes[innerLessons?.[0]?.number ?? 0]?.end}`
+                }
+                data-number={innerLessons?.[0]?.number}
+              />
+              <div className={styles.cells}>
+                {innerLessons?.map((lesson) =>
+                  <TimetableLessonInfo
+                    key={generateId(lesson)}
+                    lesson={lesson} cellSubgroup={cellSubgroup} isMerged={isMerged}
+                />)}
+              </div>
+            </LessonContainer>
+          )
+        : !isListView &&
+          <td className={classes(styles.empty, active && styles.active, isAfterEmpty && styles.border)}>
+            <div></div>
+          </td>
+      }
     </>
 
+  );
+};
+
+type ReactComponentProps = React.ButtonHTMLAttributes<HTMLElement> & { children: ReactNode };
+
+const LessonContainer: FC<ReactComponentProps & { isList: boolean }> = ({ isList, children, ...props }) => {
+  return (
+    isList
+      ? <li {...props}>{children}</li>
+      : <td {...props}>{children}</td>
   );
 };
